+++ Source/Lime/Internal/Hit.hs
@@ -10,14 +10,14 @@ import Linear
 import System.Random (StdGen)
 
 data HitData = HitData
-  { normal :: V3 Float
-  , point :: V3 Float
-  , param :: Float
+  { normal :: V3 Double
+  , point :: V3 Double
+  , param :: Double
   , -- , material :: Material
     coords :: TextureCoords
   }
 
-type Material = V3 Float -> (Color Double, Maybe Ray)
+type Material = V3 Double -> (Color Double, Maybe Ray)
 
 -- type Material =
 --   Map String (Image PixelRGBF)
@@ -26,4 +26,4 @@ type Material = V3 Float -> (Color Double, Maybe Ray)
 --   -> StdGen -- V3 Double
 --   -> (Color Double, Maybe Ray)
 
-type TextureCoords = (Float, Float)
+type TextureCoords = (Double, Double)
